weather_data_collectors/requirements.txt
requests==2.31.0
boto3==1.34.11
python-dateutil==2.8.2
tenacity==8.2.3 